+++ src/commands/install.ts
@@ -13,7 +13,7 @@ import { userError } from '../common/error';
 import { LogCallback } from '../common/log';
 import { installProvider } from '../logic/configure';
 import { initSuperface } from '../logic/init';
-import { detectSuperJson, installProfiles } from '../logic/install';
+import { detectSuperJson, installProfiles, LocalRequest as LocalInstallRequest, StoreRequest as StoreInstallRequest } from '../logic/install';
 
 const parseProviders = (
   providers?: string[],
@@ -61,6 +61,12 @@ export default class Install extends Command {
         'When set to true and when profile exists in local filesystem, overwrites them.',
       default: false,
     }),
+    local: flags.boolean({
+      char: 'l',
+      description:
+        'When set to true, profile id argument is used as a filepath to profile.supr file',
+      default: false,
+    }),
     scan: flags.integer({
       char: 's',
       description:
@@ -74,9 +80,10 @@ export default class Install extends Command {
     '$ superface install --provider twillio',
     '$ superface install sms/service@1.0',
     '$ superface install sms/service@1.0 -p twillio',
+    '$ superface install --local sms/service.supr'
   ];
 
-  private warnCallback? = (message: string) => this.log(yellow(message));
+  private warnCallback? = (message: string) => this.log('⚠️  ' + yellow(message));
   private logCallback? = (message: string) => this.log(grey(message));
 
   async run(): Promise<void> {
@@ -95,7 +102,6 @@ export default class Install extends Command {
     }
 
     let superPath = await detectSuperJson(process.cwd(), flags.scan);
-
     if (!superPath) {
       this.warnCallback?.("File 'super.json' has not been found.");
 
@@ -128,11 +134,38 @@ export default class Install extends Command {
         META_FILE
       )}'`
     );
-    await installProfiles(superPath, args.profileId, providers, {
-      logCb: this.logCallback,
-      warnCb: this.warnCallback,
-      force: flags.force,
-    });
+    
+    let installRequests: (LocalInstallRequest | StoreInstallRequest)[] = [];
+    if (args.profileId !== undefined) {
+      if (flags.local) {
+        installRequests.push(
+          {
+            kind: 'local',
+            path: args.profileId
+          }
+        );
+      } else {
+        const [profileId, version] = args.profileId.split('@');
+
+        installRequests.push(
+          {
+            kind: 'store',
+            profileId,
+            version
+          }
+        );
+      }
+    };
+
+    await installProfiles(
+      superPath,
+      installRequests,
+      {
+        logCb: this.logCallback,
+        warnCb: this.warnCallback,
+        force: flags.force
+      }
+    );
 
     this.logCallback?.(`\n\nConfiguring providers`);
     for (const providerName of providers) {