+++ src/common/format.ts
@@ -1,0 +1,8 @@
+export function formatWordPlurality(num: number, word: string): string {
+	if (num === 1) {
+	  return `${num} ${word}`;
+	} else {
+	  return `${num} ${word}s`;
+	}
+  }
+