+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+### Added
+* Support for installing local files with `install` command
+
+### Changed
+* Refactored install logic
+
 ## [0.0.5] - 2021-03-15
 
 ### Changed